feat: Created ``get_all_predictions`` on ``Store``
