+++ runner/store.py
@@ -1,8 +1,9 @@
-from typing import Dict, List, Union
+from typing import Dict, List, Union, Iterable, Tuple
 
 import pandas as pd
 
 from utils.random import random_string
+import numpy as np
 
 
 class Store:
@@ -37,3 +38,32 @@ class Store:
             df[key] = value
 
         return df
+
+    def get_all_predictions(self) -> dict:
+        outputs = dict()
+        for block_name, output in self.data.items():
+            if isinstance(output, (pd.Series, np.ndarray)):
+                output = output.tolist()
+
+            if isinstance(output, Iterable):
+                if (
+                    isinstance(output[0], Iterable)
+                    and isinstance(output[0], str) is False
+                ):
+                    predictions, probabilities = self.data_to_preds_probs(output)
+                    output = predictions
+
+                if not isinstance(output[0], str):
+                    outputs[block_name] = output
+
+        return outputs
+
+    @staticmethod
+    def data_to_preds_probs(
+        final_output: List[Union[List, Tuple]]
+    ) -> Tuple[Union[int, float]]:
+
+        predictions = [output[0] for output in final_output]
+        probabilities = [output[1] for output in final_output]
+
+        return predictions, probabilities
